requirements.txt
Django==5.2.7
psycopg2-binary==2.9.10
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
Pillow==10.4.0
